+++ NEWS.md
@@ -10,6 +10,7 @@
 - added `check_prop()` and deprecated `check_probability()`
 - added `check_grepl()` and deprecated `check_regex()` and `check_pattern()`
 - added `check_lgl()`, `check_chr()`, `check_day()`, `check_dttm()`
+- deprecated `check_flag_na()`
 
 # checkr 0.3.0